requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
PyMySQL==1.1.0
cryptography==41.0.4
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1 
pandas
xlsxwriter
pdfkit 